refactor(infra): move the migration types and table importers out of the data controller
infra-data.controller.ts was 997 lines of which only ~430 were the controller: 13 row interfaces and
a 300-line INSERT-descriptor table sat in front of two route handlers. Neither block touches
instance state, and no file outside the controller referenced any of the 18 symbols, so both move
out with a zero-width import blast radius.

The descriptor array gains a load-time completeness check. importData casts its counts object with
`as TableCounts`, so a descriptor dropped in transcription would compile, silently skip its table,
and disappear from the restored-row total that guards against wiping a database.

Pure code motion: the 29 existing controller cases pass unedited, and the restore order — FK-safe
and also the counts key order — is unchanged.
